+++ src/cache/inmemory/__tests__/cache.ts
@@ -2,7 +2,7 @@ import gql, { disableFragmentWarnings } from 'graphql-tag';
 
 import { stripSymbols } from '../../../utilities/testing/stripSymbols';
 import { cloneDeep } from '../../../utilities/common/cloneDeep';
-import { makeReference } from '../../../core';
+import { makeReference, Reference } from '../../../core';
 import { InMemoryCache, InMemoryCacheConfig } from '../inMemoryCache';
 
 disableFragmentWarnings();
@@ -1539,6 +1539,766 @@ describe("InMemoryCache#broadcastWatches", function () {
       received3,
       received4,
     ]);
+  });
+});
+
+describe("InMemoryCache#modify", () => {
+  it("should work with single modifier function", () => {
+    const cache = new InMemoryCache;
+    const query = gql`
+      query {
+        a
+        b
+        c
+      }
+    `;
+
+    cache.writeQuery({
+      query,
+      data: {
+        a: 0,
+        b: 0,
+        c: 0,
+      },
+    });
+
+    const resultBeforeModify = cache.readQuery({ query });
+    expect(resultBeforeModify).toEqual({ a: 0, b: 0, c: 0 });
+
+    cache.modify("ROOT_QUERY", (value, { fieldName }) => {
+      switch (fieldName) {
+        case "a": return value + 1;
+        case "b": return value - 1;
+        default: return value;
+      }
+    });
+
+    expect(cache.extract()).toEqual({
+      ROOT_QUERY: {
+        __typename: "Query",
+        a: 1,
+        b: -1,
+        c: 0,
+      },
+    });
+
+    const resultAfterModify = cache.readQuery({ query });
+    expect(resultAfterModify).toEqual({ a: 1, b: -1, c: 0 });
+  });
+
+  it("should work with multiple modifier functions", () => {
+    const cache = new InMemoryCache;
+    const query = gql`
+      query {
+        a
+        b
+        c
+      }
+    `;
+
+    cache.writeQuery({
+      query,
+      data: {
+        a: 0,
+        b: 0,
+        c: 0,
+      },
+    });
+
+    const resultBeforeModify = cache.readQuery({ query });
+    expect(resultBeforeModify).toEqual({ a: 0, b: 0, c: 0 });
+
+    let checkedTypename = false;
+    cache.modify("ROOT_QUERY", {
+      a(value) { return value + 1 },
+      b(value) { return value - 1 },
+      __typename(t: string, { readField }) {
+        expect(t).toBe("Query");
+        expect(readField("c")).toBe(0);
+        checkedTypename = true;
+        return t;
+      },
+    });
+    expect(checkedTypename).toBe(true);
+
+    expect(cache.extract()).toEqual({
+      ROOT_QUERY: {
+        __typename: "Query",
+        a: 1,
+        b: -1,
+        c: 0,
+      },
+    });
+
+    const resultAfterModify = cache.readQuery({ query });
+    expect(resultAfterModify).toEqual({ a: 1, b: -1, c: 0 });
+  });
+
+  it("should allow deletion using details.DELETE", () => {
+    const cache = new InMemoryCache({
+      typePolicies: {
+        Book: {
+          keyFields: ["isbn"],
+        },
+        Author: {
+          keyFields: ["name"],
+        },
+      },
+    });
+
+    const query = gql`
+      query {
+        currentlyReading {
+          title
+          isbn
+          author {
+            name
+            yearOfBirth
+          }
+        }
+      }
+    `;
+
+    const currentlyReading = {
+      __typename: "Book",
+      isbn: "147670032X",
+      title: "Why We're Polarized",
+      author: {
+        __typename: "Author",
+        name: "Ezra Klein",
+        yearOfBirth: 1983,
+      },
+    };
+
+    cache.writeQuery({
+      query,
+      data: {
+        currentlyReading,
+      }
+    });
+
+    expect(cache.extract()).toEqual({
+      ROOT_QUERY: {
+        __typename: "Query",
+        currentlyReading: {
+          __ref: 'Book:{"isbn":"147670032X"}',
+        },
+      },
+      'Book:{"isbn":"147670032X"}': {
+        __typename: "Book",
+        isbn: "147670032X",
+        author: {
+          __ref: 'Author:{"name":"Ezra Klein"}',
+        },
+        title: "Why We're Polarized",
+      },
+      'Author:{"name":"Ezra Klein"}': {
+        __typename: "Author",
+        name: "Ezra Klein",
+        yearOfBirth: 1983,
+      },
+    });
+
+    const authorId = cache.identify(currentlyReading.author);
+    expect(authorId).toBe('Author:{"name":"Ezra Klein"}');
+
+    cache.modify(authorId, {
+      yearOfBirth(yob) {
+        return yob + 1;
+      },
+    });
+
+    const yobResult = cache.readFragment({
+      id: authorId,
+      fragment: gql`fragment YOB on Author { yearOfBirth }`,
+    });
+
+    expect(yobResult).toEqual({
+      __typename: "Author",
+      yearOfBirth: 1984,
+    });
+
+    const bookId = cache.identify(currentlyReading);
+
+    // Modifying the Book in order to modify the Author is fancier than
+    // necessary, but we want fancy use cases to work, too.
+    cache.modify(bookId, {
+      author(author: Reference, { readField }) {
+        expect(readField("title")).toBe("Why We're Polarized");
+        expect(readField("name", author)).toBe("Ezra Klein");
+        cache.modify(cache.identify({
+          __typename: readField("__typename", author),
+          name: readField("name", author),
+        }), {
+          yearOfBirth(yob, { DELETE }) {
+            expect(yob).toBe(1984);
+            return DELETE;
+          },
+        });
+        return author;
+      }
+    });
+
+    const snapshotWithoutYOB = cache.extract();
+    expect(snapshotWithoutYOB[authorId].yearOfBirth).toBeUndefined();
+    expect("yearOfBirth" in snapshotWithoutYOB[authorId]).toBe(false);
+    expect(snapshotWithoutYOB).toEqual({
+      ROOT_QUERY: {
+        __typename: "Query",
+        currentlyReading: {
+          __ref: 'Book:{"isbn":"147670032X"}',
+        },
+      },
+      'Book:{"isbn":"147670032X"}': {
+        __typename: "Book",
+        isbn: "147670032X",
+        author: {
+          __ref: 'Author:{"name":"Ezra Klein"}',
+        },
+        title: "Why We're Polarized",
+      },
+      'Author:{"name":"Ezra Klein"}': {
+        __typename: "Author",
+        name: "Ezra Klein",
+        // yearOfBirth is gone now
+      },
+    });
+
+    // Delete the whole Book.
+    cache.modify(bookId, (_, { DELETE }) => DELETE);
+
+    const snapshotWithoutBook = cache.extract();
+    expect(snapshotWithoutBook[bookId]).toBeUndefined();
+    expect(bookId in snapshotWithoutBook).toBe(false);
+    expect(snapshotWithoutBook).toEqual({
+      ROOT_QUERY: {
+        __typename: "Query",
+        currentlyReading: {
+          __ref: 'Book:{"isbn":"147670032X"}',
+        },
+      },
+      'Author:{"name":"Ezra Klein"}': {
+        __typename: "Author",
+        name: "Ezra Klein",
+      },
+    });
+
+    // Delete all fields of the Author, which also removes the object.
+    cache.modify(authorId, {
+      __typename(_, { DELETE }) { return DELETE },
+      name(_, { DELETE }) { return DELETE },
+    });
+
+    const snapshotWithoutAuthor = cache.extract();
+    expect(snapshotWithoutAuthor[authorId]).toBeUndefined();
+    expect(authorId in snapshotWithoutAuthor).toBe(false);
+    expect(snapshotWithoutAuthor).toEqual({
+      ROOT_QUERY: {
+        __typename: "Query",
+        currentlyReading: {
+          __ref: 'Book:{"isbn":"147670032X"}',
+        },
+      },
+    });
+
+    cache.modify("ROOT_QUERY", (_, { DELETE }) => DELETE);
+    expect(cache.extract()).toEqual({});
+  });
+
+  it("can remove specific items from paginated lists", () => {
+    const cache = new InMemoryCache({
+      typePolicies: {
+        Thread: {
+          keyFields: ["tid"],
+
+          fields: {
+            comments: {
+              merge(existing: Reference[], incoming: Reference[], { args, mergeObjects }) {
+                const merged = existing ? existing.slice(0) : [];
+                const end = args.offset + Math.min(args.limit, incoming.length);
+                for (let i = args.offset; i < end; ++i) {
+                  merged[i] = mergeObjects(merged[i], incoming[i - args.offset]);
+                }
+                return merged;
+              },
+
+              read(existing: Reference[], { args }) {
+                const page = existing && existing.slice(
+                  args.offset,
+                  args.offset + args.limit,
+                );
+                if (page && page.length > 0) {
+                  return page;
+                }
+              },
+            },
+          },
+        },
+
+        Comment: {
+          keyFields: ["id"],
+        },
+      },
+    });
+
+    const query = gql`
+      query GetThread($offset: Int, $limit: Int) {
+        thread {
+          tid
+          comments(offset: $offset, limit: $limit) {
+            id
+            text
+          }
+        }
+      }
+    `;
+
+    cache.writeQuery({
+      query,
+      data: {
+        thread: {
+          __typename: "Thread",
+          tid: 123,
+          comments: [{
+            __typename: "Comment",
+            id: "c1",
+            text: "first post",
+          }, {
+            __typename: "Comment",
+            id: "c2",
+            text: "I have thoughts",
+          }, {
+            __typename: "Comment",
+            id: "c3",
+            text: "friendly ping",
+          }]
+        }
+      },
+      variables: {
+        offset: 0,
+        limit: 3,
+      },
+    });
+
+    expect(cache.extract()).toEqual({
+      ROOT_QUERY: {
+        __typename: "Query",
+        thread: {
+          __ref: 'Thread:{"tid":123}',
+        },
+      },
+      'Thread:{"tid":123}': {
+        __typename: "Thread",
+        tid: 123,
+        comments: [
+          { __ref: 'Comment:{"id":"c1"}' },
+          { __ref: 'Comment:{"id":"c2"}' },
+          { __ref: 'Comment:{"id":"c3"}' },
+        ],
+      },
+      'Comment:{"id":"c1"}': {
+        __typename: "Comment",
+        id: "c1",
+        text: "first post",
+      },
+      'Comment:{"id":"c2"}': {
+        __typename: "Comment",
+        id: "c2",
+        text: "I have thoughts",
+      },
+      'Comment:{"id":"c3"}': {
+        __typename: "Comment",
+        id: "c3",
+        text: "friendly ping",
+      },
+    });
+
+    cache.modify(cache.identify({
+      __typename: "Thread",
+      tid: 123,
+    }), {
+      comments(comments: Reference[], { readField }) {
+        debugger;
+        expect(Object.isFrozen(comments)).toBe(true);
+        expect(comments.length).toBe(3);
+        const filtered = comments.filter(comment => {
+          return readField("id", comment) !== "c1";
+        });
+        expect(filtered.length).toBe(2);
+        return filtered;
+      },
+    });
+
+    expect(cache.gc()).toEqual(['Comment:{"id":"c1"}']);
+
+    expect(cache.extract()).toEqual({
+      ROOT_QUERY: {
+        __typename: "Query",
+        thread: {
+          __ref: 'Thread:{"tid":123}',
+        },
+      },
+      'Thread:{"tid":123}': {
+        __typename: "Thread",
+        tid: 123,
+        comments: [
+          { __ref: 'Comment:{"id":"c2"}' },
+          { __ref: 'Comment:{"id":"c3"}' },
+        ],
+      },
+      'Comment:{"id":"c2"}': {
+        __typename: "Comment",
+        id: "c2",
+        text: "I have thoughts",
+      },
+      'Comment:{"id":"c3"}': {
+        __typename: "Comment",
+        id: "c3",
+        text: "friendly ping",
+      },
+    });
+  });
+
+  it("should not revisit deleted fields", () => {
+    const cache = new InMemoryCache;
+    const query = gql`query { a b c }`;
+
+    cache.recordOptimisticTransaction(cache => {
+      cache.writeQuery({
+        query,
+        data: {
+          a: 1,
+          b: 2,
+          c: 3,
+        },
+      })
+    }, "transaction");
+
+    cache.modify("ROOT_QUERY", {
+      b(value, { DELETE }) {
+        expect(value).toBe(2);
+        return DELETE;
+      },
+    }, true);
+
+    expect(cache.extract(true)).toEqual({
+      ROOT_QUERY: {
+        __typename: "Query",
+        a: 1,
+        c: 3,
+      },
+    });
+
+    cache.modify("ROOT_QUERY", (value, { fieldName }) => {
+      expect(fieldName).not.toBe("b");
+      if (fieldName === "a") expect(value).toBe(1);
+      if (fieldName === "c") expect(value).toBe(3);
+    }, true);
+
+    cache.removeOptimistic("transaction");
+
+    expect(cache.extract(true)).toEqual({});
+  });
+
+  it("should broadcast watches for queries with changed fields", () => {
+    const cache = new InMemoryCache;
+    const queryA = gql`{ a { value } }`;
+    const queryB = gql`{ b { value } }`;
+
+    cache.writeQuery({
+      query: queryA,
+      data: {
+        a: {
+          __typename: "A",
+          id: 1,
+          value: 123,
+        },
+      }
+    });
+
+    cache.writeQuery({
+      query: queryB,
+      data: {
+        b: {
+          __typename: "B",
+          id: 1,
+          value: 321,
+        },
+      }
+    });
+
+    expect(cache.extract()).toEqual({
+      ROOT_QUERY: {
+        __typename: "Query",
+        a: { __ref: "A:1" },
+        b: { __ref: "B:1" },
+      },
+      "A:1": {
+        __typename: "A",
+        value: 123,
+      },
+      "B:1": {
+        __typename: "B",
+        value: 321,
+      },
+    });
+
+    let aResults = [];
+    cache.watch({
+      query: queryA,
+      optimistic: true,
+      immediate: true,
+      callback(data) {
+        aResults.push(data);
+      },
+    });
+
+    let bResults = [];
+    cache.watch({
+      query: queryB,
+      optimistic: true,
+      immediate: true,
+      callback(data) {
+        bResults.push(data);
+      },
+    });
+
+    function makeResult(
+      __typename: string,
+      value: number,
+      complete: boolean = true,
+    ) {
+      return {
+        complete,
+        result: {
+          [__typename.toLowerCase()]: {
+            __typename,
+            value,
+          },
+        },
+      };
+    }
+
+    const a123 = makeResult("A", 123);
+    const b321 = makeResult("B", 321);
+
+    expect(aResults).toEqual([a123]);
+    expect(bResults).toEqual([b321]);
+
+    const aId = cache.identify({ __typename: "A", id: 1 });
+    const bId = cache.identify({ __typename: "B", id: 1 });
+
+    cache.modify(aId, {
+      value(x: number) {
+        return x + 1;
+      },
+    });
+
+    const a124 = makeResult("A", 124);
+
+    expect(aResults).toEqual([a123, a124]);
+    expect(bResults).toEqual([b321]);
+
+    cache.modify(bId, {
+      value(x: number) {
+        return x + 1;
+      },
+    });
+
+    const b322 = makeResult("B", 322);
+
+    expect(aResults).toEqual([a123, a124]);
+    expect(bResults).toEqual([b321, b322]);
+  });
+
+  it("should handle argument-determined field identities", () => {
+    const cache = new InMemoryCache({
+      typePolicies: {
+        Query: {
+          fields: {
+            book: {
+              keyArgs: ["isbn"],
+            },
+          },
+        },
+        Book: {
+          keyFields: ["isbn"],
+        },
+      },
+    });
+
+    function addBook(isbn: string, title: string) {
+      cache.writeQuery({
+        query: gql`
+          query {
+            book(isbn: $isbn) {
+              isbn
+              title
+            }
+          }
+        `,
+        data: {
+          book: {
+            __typename: "Book",
+            isbn,
+            title,
+          },
+        },
+        variables: {
+          isbn,
+        },
+      });
+    }
+
+    addBook("147670032X", "Why We're Polarized");
+    addBook("1760641790", "How To Do Nothing");
+    addBook("0735211280", "Spineless");
+
+    const fullSnapshot = {
+      ROOT_QUERY: {
+        __typename: "Query",
+        'book:{"isbn":"0735211280"}': {
+          __ref: 'Book:{"isbn":"0735211280"}',
+        },
+        'book:{"isbn":"147670032X"}': {
+          __ref: 'Book:{"isbn":"147670032X"}',
+        },
+        'book:{"isbn":"1760641790"}': {
+          __ref: 'Book:{"isbn":"1760641790"}',
+        },
+      },
+      'Book:{"isbn":"147670032X"}': {
+        __typename: "Book",
+        isbn: "147670032X",
+        title: "Why We're Polarized",
+      },
+      'Book:{"isbn":"1760641790"}': {
+        __typename: "Book",
+        isbn: "1760641790",
+        title: "How To Do Nothing",
+      },
+      'Book:{"isbn":"0735211280"}': {
+        __typename: "Book",
+        isbn: "0735211280",
+        title: "Spineless",
+      },
+    };
+
+    expect(cache.extract()).toEqual(fullSnapshot);
+
+    function check(isbnToDelete?: string) {
+      let bookCount = 0;
+
+      cache.modify("ROOT_QUERY", {
+        book(book: Reference, {
+          fieldName,
+          storeFieldName,
+          isReference,
+          readField,
+          DELETE,
+        }) {
+          expect(fieldName).toBe("book");
+          expect(isReference(book)).toBe(true);
+          expect(typeof readField("title", book)).toBe("string");
+          expect(readField("__typename", book)).toBe("Book");
+
+          const parts = storeFieldName.split(":");
+          expect(parts.shift()).toBe("book");
+          const keyArgs = JSON.parse(parts.join(":"));
+          expect(typeof keyArgs.isbn).toBe("string");
+          expect(Object.keys(keyArgs)).toEqual(["isbn"]);
+
+          expect(readField("isbn", book)).toBe(keyArgs.isbn);
+
+          if (isbnToDelete === keyArgs.isbn) {
+            return DELETE;
+          }
+
+          ++bookCount;
+
+          return book;
+        },
+      });
+
+      return bookCount;
+    }
+
+    // No change from repeatedly calling check().
+    expect(check()).toBe(3);
+    expect(check()).toBe(3);
+
+    expect(check("0735211280")).toBe(2);
+    expect(check("147670032X")).toBe(1);
+
+    // No change from re-deleting already-deleted ISBNs.
+    expect(check("0735211280")).toBe(1);
+    expect(check("147670032X")).toBe(1);
+
+    expect(cache.extract()).toEqual({
+      ROOT_QUERY: {
+        __typename: "Query",
+        'book:{"isbn":"1760641790"}': {
+          __ref: 'Book:{"isbn":"1760641790"}',
+        },
+      },
+      'Book:{"isbn":"147670032X"}': {
+        __typename: "Book",
+        isbn: "147670032X",
+        title: "Why We're Polarized",
+      },
+      'Book:{"isbn":"1760641790"}': {
+        __typename: "Book",
+        isbn: "1760641790",
+        title: "How To Do Nothing",
+      },
+      'Book:{"isbn":"0735211280"}': {
+        __typename: "Book",
+        isbn: "0735211280",
+        title: "Spineless",
+      },
+    });
+
+    expect(cache.gc().sort()).toEqual([
+      'Book:{"isbn":"0735211280"}',
+      'Book:{"isbn":"147670032X"}',
+    ]);
+
+    expect(cache.extract()).toEqual({
+      ROOT_QUERY: {
+        __typename: "Query",
+        'book:{"isbn":"1760641790"}': {
+          __ref: 'Book:{"isbn":"1760641790"}',
+        },
+      },
+      'Book:{"isbn":"1760641790"}': {
+        __typename: "Book",
+        isbn: "1760641790",
+        title: "How To Do Nothing",
+      },
+    });
+
+    expect(check("1760641790")).toBe(0);
+
+    expect(cache.extract()).toEqual({
+      ROOT_QUERY: {
+        __typename: "Query",
+      },
+      'Book:{"isbn":"1760641790"}': {
+        __typename: "Book",
+        isbn: "1760641790",
+        title: "How To Do Nothing",
+      },
+    });
+
+    expect(cache.gc()).toEqual([
+      'Book:{"isbn":"1760641790"}',
+    ]);
+
+    expect(cache.extract()).toEqual({
+      ROOT_QUERY: {
+        __typename: "Query",
+      },
+    });
   });
 });