+++ src/cache/inmemory/types.ts
@@ -2,6 +2,7 @@ import { DocumentNode } from 'graphql';
 
 import { Transaction } from '../core/cache';
 import { StoreValue } from '../../utilities/graphql/storeUtils';
+import { Modifiers, Modifier, FieldValueGetter } from './entityStore';
 export { StoreValue }
 
 export interface IdGetterObj extends Object {
@@ -22,6 +23,7 @@ export interface NormalizedCache {
   has(dataId: string): boolean;
   get(dataId: string, fieldName: string): StoreValue;
   merge(dataId: string, incoming: StoreObject): void;
+  modify(dataId: string, modifiers: Modifier<any> | Modifiers): boolean;
   delete(dataId: string, fieldName?: string): boolean;
   clear(): void;
 
@@ -45,6 +47,8 @@ export interface NormalizedCache {
    */
   retain(rootId: string): number;
   release(rootId: string): number;
+
+  getFieldValue: FieldValueGetter;
 }
 
 /**